Shift manual queries in data/sql into respective RQ folders + update CODEOWNERS
