+++ apps/studio/components/interfaces/Storage/StorageSettings/StorageFileSizeLimitErrorMessage.tsx
@@ -1,9 +1,9 @@
+import { InlineLink } from 'components/ui/InlineLink'
+import { LARGEST_SIZE_LIMIT_BUCKETS_COUNT } from 'data/storage/storage.sql'
 import Link from 'next/link'
 import { type FieldError } from 'react-hook-form'
-
-import { InlineLink } from 'components/ui/InlineLink'
-import { LARGEST_SIZE_LIMIT_BUCKETS_COUNT } from 'data/sql/queries/get-largest-size-limit-buckets'
 import { cn, Tooltip, TooltipContent, TooltipTrigger } from 'ui'
+
 import {
   decodeBucketLimitErrorMessage,
   formatBytesForDisplay,
