+++ apps/studio/data/database-cron-jobs/database-cron-jobs-minimal-infinite-query.ts
@@ -2,12 +2,12 @@ import { InfiniteData, useInfiniteQuery } from '@tanstack/react-query'
 import { executeSql } from 'data/sql/execute-sql-query'
 import type { ResponseError, UseCustomInfiniteQueryOptions } from 'types'
 
-import { getCronJobsMinimalSql } from '../sql/queries/get-cron-jobs'
 import {
   CRON_JOBS_PAGE_LIMIT,
   CronJob,
   DatabaseCronJobRunsVariables,
 } from './database-cron-jobs-infinite-query'
+import { getCronJobsMinimalSql } from './database-cron-jobs.sql'
 import { databaseCronJobsKeys } from './keys'
 
 export async function getDatabaseCronJobsMinimal({
