+++ apps/studio/data/database-extensions/database-extensions.sql.ts
