+++ apps/studio/data/database-indexes/database-indexes.sql.ts
