+++ apps/studio/data/database/database.sql.ts
@@ -1,5 +1,6 @@
 import { literal } from '@supabase/pg-meta/src/pg-format'
-import { sqlKeys } from '../keys'
+
+import { sqlKeys } from '../sql/keys'
 
 export const getLiveTupleEstimate = (table: string, schema: string = 'public') => {
   const sql = /* SQL */ `
